requirements.txt
cryptography
argon2-cffi
blake3